refactor(core.views.py): refactoring code : add files and functions
