+++ django/ft_transcendence/chat/models.py
@@ -35,3 +35,6 @@ class Message(models.Model):
 
     def __str__(self):
         return u'%s - %d' % (self.refChat,self.type)
+
+    def last_10_messages(chat):
+        return Message.objects.filter(refChat=chat).order_by('-createdAt')[:10]
