+++ django/ft_transcendence/chat/urls.py
@@ -2,8 +2,10 @@ from django.urls import path
 
 from . import views
 
+app_name = 'chat'
+
 urlpatterns = [
     path("", views.index, name="index"),
     path("<str:room_name>/<int:userID>", views.room, name="room"),
-    path("send_direct_message/<int:userID>/", views.send_direct_message, name='send_direct_message')
+    path("direct_message/", views.direct_message, name='direct_message')
 ]
