+++ django/ft_transcendence/core/profile_utils.py
@@ -13,8 +13,27 @@ def get_or_create_chat(request, chatWithUser):
 		).first()
 
 	if not chat:
-		print("No chat yet")
 		chat = Chat.objects.create(fromUser=request.user, toUser=chatWithUser)
 
-	print("Chat exists")
 	return chat
+
+def _has_friend_request_(request, user_profile):
+	friend_request_receiver = request.user.receiver.all()
+	friend_request_sender   = request.user.sender.all()
+	# Is there friend request that the user has received from user_profile
+	for received in friend_request_receiver:
+		if received.sender == user_profile:
+			return True
+	# Is there friend request that the user has send to user_profile
+	for send in friend_request_sender:
+		if send.receiver == user_profile:
+			return True
+	return False
+
+def _is_friend_(request, user_profile):
+	# Is userprofile a friend
+	for friend in request.user.userprofile.friends.all():
+		if friend.id == user_profile.id:
+			return True
+	return False
+
