+++ django/ft_transcendence/core/templates/core/500.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Server Error (500)</title>
+</head>
+<body>
+    <h1>Oops! Something went wrong.</h1>
+    <p>We're sorry, but an internal server error has occurred.</p>
+</body>
+</html>
