+++ django/ft_transcendence/core/templates/core/sidenav.html
@@ -15,6 +15,7 @@
 	<a href="/home/">Home</a>
 	<a href="/my_profile/">Profile</a>
 	<a href="/social/">Social</a>
+	<a href=" {% url 'chat:direct_message' %}">Chat</a>
 	<a href="#">Game Stats</a>
 	<a href="/logout/">Logout</a>
 </div>
